requirementsTerceros.txt
aiopulsar-py==0.1.0a0
anyio==3.6.2
asgiref==3.6.0
attrs==22.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
coverage==6.5.0
cryptography==39.0.2
Faker==17.6.0
fastapi==0.89.1
fastavro==0.24.0
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
flask-swagger==0.2.14
graphql-core==3.2.3
greenlet==2.0.1
h11==0.14.0
httptools==0.5.0
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.32
packaging==21.3
pluggy==1.0.0
protobuf==3.20.3
pulsar-client==2.10.2
py==1.11.0
pycparser==2.21
pydantic==1.10.6
PyDispatcher==2.0.6
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.5
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
sse-starlette==1.2.1
starlette==0.22.0
strawberry-graphql==0.156.3
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
Werkzeug==2.2.2
